TIP-870: Move pdf controller to Akeneo\Enrichement
